Log Context Processor
Create a log message processor to add static context to log messages.
The context model used to gather data still needs refinement but
functions at a basic level.

Create a virtual type for a logger instance with an added processor to
add the context to log messages. Inject address types that require a
logger with the virtual type.