US36651 Functional and Integration Behat Tests
Add Behat as a dev dependency for testing the extension. Define tests
for address validation in the address book using Behat.

Reorganize tests based on the type of test: unit, functional,
integration.

Updates to the integration environment. Installs Magento 2 and the ROM
extension within a Docker container. Environment will no longer be baked
into the dev environment, nor will it be mounted from a host to the
Docker container - improves performance of the environment.